# contracts/center-api.openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Monitoring Center API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /settings/users:
    get:
      operationId: listUsers
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200':
          description: Administrator user list without credentials
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserWrite'}
      responses:
        '201': {description: User created}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /settings/users/{username}:
    put:
      operationId: updateUser
      parameters:
        - name: username
          in: path
          required: true
          schema: {type: string, maxLength: 80}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserWrite'}
      responses:
        '204': {description: User updated}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: JSESSIONID}
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: CSRF token missing or access denied
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Identifier:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{1,63}$'
...
    PageMetadata:
      type: object
      required: [page, size, total]
      properties:
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
...
    ManagedUser:
      type: object
      required: [username, role, enabled, project_ids, created_at, updated_at]
      properties:
        username: {type: string}
        role: {type: string, enum: [ADMIN, OPERATOR, VIEWER]}
        enabled: {type: boolean}
        project_ids: {type: array, items: {type: string}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    UserWrite:
      type: object
      required: [username, role, enabled, projectIds]
      properties:
        username: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._@-]{1,79}$'}
        password: {type: string, format: password, writeOnly: true, description: Required on creation; blank on update preserves existing password. Supplied passwords need at least 12 characters and at most 72 UTF-8 bytes.}
        role: {type: string, enum: [ADMIN, OPERATOR, VIEWER]}
        enabled: {type: boolean}
        projectIds: {type: array, maxItems: 200, uniqueItems: true, items: {$ref: '#/components/schemas/Identifier'}}
    UserPage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/ManagedUser'}}
...
    Error:
      type: object
      required: [code]
      properties:
        code: {type: string}
...